Feat: Refactoring, login page, admin home page start
