+++ clubhub/src/database.py
@@ -1,0 +1,2 @@
+import util
+db = util.open_database(config["DATABASE_URL"])
