+++ clubhub/src/globals.py
@@ -1,0 +1,4 @@
+import util
+
+config = util.load_configuration()
+db = util.open_database(config["DATABASE_URL"])
