+++ clubhub/src/templates/admin/home.html
@@ -1,0 +1,25 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <title>Admin Home | ClubHub</title>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <link href="static/styles/index.css" rel="stylesheet">
+    <script src="https://kit.fontawesome.com/c2b7a7d1d7.js" crossorigin="anonymous"></script>
+  </head>
+  <body>
+
+    <main class="app__main">
+      {% set page = 'users' %}
+      {% include 'admin/nav.html' %}
+
+      <div class="app__content">
+          <h1>
+            Admin Users
+           </h1>
+      </div>
+
+    </main>
+
+  </body>
+</html>
