+++ clubhub/src/templates/admin/nav.html
@@ -1,0 +1,36 @@
+<!-- Side navigation component for admin -->
+
+
+
+<nav class="navbar">
+    <div class="app_logo_container">
+        <h3>ClubHub</h3>
+        <p>Administrator</p>
+    </div>
+    <div class="nav__links_container">
+        <ul class="nav__links">
+            {% if page == 'users'%}
+            <li class="nav__link_selected"><a href="/admin/users">Users <i class="fa-solid fa-users"></i></a></li>
+            {% else %}
+            <li><a href="/admin/users">Users <i class="fa-solid fa-users"></i></a></li>
+            {% endif %}
+
+            {% if page == 'clubs' %}
+            <li class="nav__link_selected"><a href="/admin/clubs">Clubs <i class="fa-solid fa-volleyball"></i></a></li>
+            {% else %}
+            <li><a href="/admin/clubs">Clubs <i class="fa-solid fa-volleyball"></i></a></li>
+            {% endif %}
+
+            {% if page == 'events' %}
+            <li class="nav__link_selected"><a href="/admin/events">Events <i class="fa-solid fa-calendar-days"></i></a></li>
+            {% else %}
+            <li><a href="/admin/events">Events <i class="fa-solid fa-calendar-days"></i></a></li>
+            {% endif %}
+
+        </ul>
+    </div>
+
+    <div class="signout_container">
+
+    </div>
+</nav>
