+++ clubhub/src/templates/navigation.html
@@ -1,0 +1,1 @@
+<!-- Re-usable nav wrapper to use on pages -->
